feat : 팝업 전환 구현
